+++ src/hdmf/utils.py
@@ -894,6 +894,7 @@ def _unwrap_scalar(value):
     return value
 
 
+
 def pystr(s):
     """
     Convert a string of characters to Python str object